+++ muttlib/dbconn/bigquery.py
@@ -16,7 +16,7 @@ except ModuleNotFoundError:
     logger.debug("No BigQuery support.")
 
 
-BIGQUERY_DB_TYPE = 'bigquery'
+BIGQUERY_DB_TYPE = "bigquery"
 
 
 class BigQueryClient(BaseClient):
@@ -58,7 +58,9 @@ class BigQueryClient(BaseClient):
         return f"{self.project}.{self.db}.{self.table}"
 
     def _read_cred(
-        self, auth: Optional[str], auth_file: Optional[Union[str, Path]],
+        self,
+        auth: Optional[str],
+        auth_file: Optional[Union[str, Path]],
     ):
         """Create valid OAuth2 credentials for bigquery.
 
@@ -212,5 +214,5 @@ class BigQueryClient(BaseClient):
         try:
             connection.get_table(table)  # Make an API request.
         except exceptions.NotFound:
-            self.execute(sql, params={'table_id': table}, connection=connection)
+            self.execute(sql, params={"table_id": table}, connection=connection)
             logger.info(f"Created {table}")