+++ muttlib/dbconn/postgres.py
@@ -4,7 +4,7 @@ from muttlib.dbconn.base import EngineBaseClient
 
 logger = logging.getLogger(__name__)
 
-POSTGRES_DB_TYPE = 'postgres'
+POSTGRES_DB_TYPE = "postgres"
 POSTGRES_DIALECT = "postgresql"
 
 try: