+++ muttlib/dbconn/sqlite.py
@@ -1,9 +1,9 @@
 from muttlib.dbconn.base import EngineBaseClient
 
-SQLITE_DB_TYPE = 'sqlite'
+SQLITE_DB_TYPE = "sqlite"
 
 
 class SqliteClient(EngineBaseClient):
     """Create SQLite DB client."""
 
-    default_dialect = 'sqlite'
+    default_dialect = "sqlite"