+++ muttlib/dbconn/sqlserver.py
@@ -1,6 +1,6 @@
 from muttlib.dbconn.base import EngineBaseClient
 
-SQLSERVER_DB_TYPE = 'sql_server'
+SQLSERVER_DB_TYPE = "sql_server"
 SQLSERVER_DIALECT = "mssql"
 
 
@@ -8,7 +8,7 @@ class SqlServerClient(EngineBaseClient):
     """SQLServer client."""
 
     default_dialect = SQLSERVER_DIALECT
-    default_driver = 'pymssql'
+    default_driver = "pymssql"
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)