+++ tests/dbconn/test_sqlite.py
@@ -11,9 +11,11 @@ def sqlite_client():
 
 
 def test_url_sqlite():
-    client = SqliteClient(database="database.db",)
+    client = SqliteClient(
+        database="database.db",
+    )
     engine = client.get_engine()
-    assert str(engine.url) == 'sqlite:///database.db'
+    assert str(engine.url) == "sqlite:///database.db"
 
 
 def test_sqlite_execute():