+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.5.0] - 2022-03-08
+
+### Fixed
+  - Updated dependencies
+  - Deprecate support for python3.7
+  - Renamed dependencie `fbprophet` -> `prophet`
 ## [1.4.15] - 2022-03-08
 
 ### Fixed